make base Agent class without discord features
